refactor(console): controller 패키지 분리
